update de tout ce que j'ai rm -rf (jsuis un bouffon)

+++ libft/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: lcroxatt <lcroxatt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/28 17:58:43 by lcroxatt          #+#    #+#             */
-/*   Updated: 2023/10/28 18:11:57 by lcroxatt         ###   ########.fr       */
+/*   Updated: 2024/05/31 16:54:05 by lcroxatt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	resultat;
-	int	signe;
+	long	resultat;
+	long	signe;
 
 	signe = 1;
 	resultat = 0;
